fix: rename and move central file
